feat: ts files instead of js
