+++ package.json
@@ -2,14 +2,13 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "index.ts",
     "engines": {
         "node": ">=20.0.0"
     },
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "tsx ./index.ts",
+        "start:dev": "nodemon",
         "test": "echo \"Error: no test specified\" && exit 1",
         "format": "prettier --write \"src/**/*.ts\"",
         "lint": "eslint \"src/**/*.ts\" --fix"
